[REM-333] refactor: update credential form resources and structure
- CredentialsForm.vue: changed import from formusr to formcred, removed unused email input
- CredentialsView.vue: updated import from formusr to formcred
- CredentialsComponent.vue: changed imports from formusr to formcred
- formcred.ts: created new resource file for credential management with reactive form state and submission logic
