+++ client/src/routes/credentials/resources/formcred.ts
@@ -1,5 +1,6 @@
+import { AxiosError, AxiosResponse } from "axios";
 import { reactive, ref } from "vue";
-import { $, api } from "../../../main";
+import { api } from "../../../main";
 export const form = reactive({
   id: 0,
   name_credential: "",
@@ -14,6 +15,9 @@ export const form = reactive({
   },
 });
 
+export const selected2 = ref(null);
+export const systems_list = ref<unknown[]>([{ value: null, text: "Carregando", disabled: true }]);
+
 export const delete_call = ref(false);
 export const current_action = ref("");
 export const state_modal = ref(false);
@@ -23,26 +27,18 @@ export const to_modal_message = ref(false);
 
 export async function submitForm(e: Event) {
   submited.value = true;
-
   e.preventDefault();
 
+  const systembot = selected2.value as unknown as string;
   const formData = new FormData();
-  formData.append("nome_usuario", form.name_credential);
+  formData.append("nome_cred", form.name_credential);
+  formData.append("system", systembot);
+  formData.append("auth_method", "pw");
   formData.append("login", form.login);
   formData.append("password", form.password);
 
-  if (current_action.value.includes("Editar")) {
-    formData.append("id", form.id.toString());
-    formData.append("method_request", "UPDATE");
-  } else if (current_action.value.includes("Cadastrar")) {
-    formData.append("method_request", "INSERT");
-  } else if (delete_call.value === true) {
-    formData.append("id", form.id.toString());
-    formData.append("method_request", "DELETE");
-  }
-
   api
-    .post("/users", formData, {
+    .post("/cadastro_credencial", formData, {
       withXSRFToken: true,
       withCredentials: true,
       headers: {
@@ -51,17 +47,17 @@ export async function submitForm(e: Event) {
         "x-csrf-token": sessionStorage.getItem("x-csrf-token") || "",
       },
     })
-    .then((response) => {
+    .then((response: AxiosResponse) => {
       if (response.status === 200) {
-        $("#message").text(response.data.message);
-      } else {
         $("#message").text(response.data.message);
       }
       to_modal_message.value = true;
     })
+    .catch((error: AxiosError) => {
+      const response = error.response as AxiosResponse;
+      const message = response.data.message as string;
 
-    .catch((response) => {
-      $("#message").text(response.data.message);
+      $("#message").text(message);
 
       to_modal_message.value = true;
     });
